+++ Phoenix/PHOptionConstants.h
@@ -1,0 +1,9 @@
+/*
+ * Phoenix is released under the MIT License. Refer to https://github.com/kasper/phoenix/blob/master/LICENSE.md
+ */
+
+@import Foundation;
+
+#pragma mark - Window Options
+
+static NSString * const PHWindowVisibilityOptionKey = @"visible";